FIX - Changed ids in sample/default data to match the id ranges mentioned in developer guide.
git-svn-id: http://scarab.tigris.org/svn/scarab@10994 3c7165bb-d2ef-0310-b724-bafd9a86081f

+++ branches/release/0.22/src/test/org/tigris/scarab/om/QueryTest.java
@@ -138,7 +138,7 @@ public class QueryTest extends BaseTurbineTestCase
     public void testSaveAndSendEmail() throws Exception
     {
         System.out.println("\ntestSaveAndSendEmail()");
-        query1.setUserId(new Integer(2));
+        query1.setUserId(new Integer(1002));
         query1.setName("Test query 2");
         query1.setValue("&searchId=2&searchisp=asc");
         query1.setDescription("Description for test query 2");
